+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Fixed
 - Method `normalizeArrayResponse` from `Serializer.Offline` did not expect that adapter can return response in various format.
 
+### Removed
+- Obsolete code in `syncer` service.
+
 ## [0.6.2-beta.13] - 2016-10-05
 ### Fixed
 - Casting to string only `Date` instead of all types.